files/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
asyncpg==0.29.0
python-dotenv==1.0.1
